Fixed Cython problems with installing module from sdist
tox creates a source distribution from the module under test and installs from that. For this to
work with cythonize, the original .pyx files have to be added to the source distribution. This is
what MANIFEST.in is for.

The initial approach using cython_ext was flawed, so it was removed. The current solution is not final. In the final version the actual distributin should not depend on Cython at all (since it includes the translated c files), and end users should not need Cython to install Mitty.