Fixed test invocation from tox
Tests are now outside of the actual source tree of mitty. This is required because otherwise the
local sources will always override the installed version of mitty. However, for tox to work as
expected, we want to explicitly test the installed version.

tox.ini has been updated to ensure that the installed version is tested, rather than the source
version.