Fixed Cython bootstraping
Cython is now only imported if present, otherwise the extension are specified manually. For
installing from source-distributions (which already include the *.c-files), Cython is no longer
required (this is relevant for end-users).

For installing from sources or for creating a working source distribution, Cython is required.

Since tox does not support installing requirements for creating the source distribution, Cython has
to be manually installed before running tox or doing a development install.