Switched all internals to 0-indexing
Convention modified such that VCFs are written with 1-indexing but internals are all kept at 0-indexing. This is much less confusing.

This only requires one change to io to change how we write out VCF files (and updating tests) and to the snp plugin to remove the index conversion.

This will make the reads code much easier to read and be less prone to index conversion bugs